Added Multiplayer support and input mode switch
Open the MenuMap level in content/menu, play from their, press create to create a server and refresh on another client to search for servers.

When in a server press 1 to switch to third person and 2 to switch to vr, press 3 when in third person to open the controller widget.